+++ src/Libraries/AFILE/Source/compose.h
@@ -19,11 +19,14 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 //		COMPOSEW.H		WIDE8 Compose buffering
 //		Rex E. Bradford (REX)
 //
-//		These routines implement a buffer for composing images.  You should
-//		use one of the following types when defining such a compose buffer:
+//		These routines implement a buffer for composing images.  You
+// should
+//		use one of the following types when defining such a compose
+// buffer:
 //
 //		BMT_FLAT8:	Compose 8-bit images, don't need difference info
-//		BMT_FLAT24:	Compose 24-bit images, don't need difference info
+//		BMT_FLAT24:	Compose 24-bit images, don't need difference
+// info
 
 /*
  * $Header: r:/prj/lib/src/afile/RCS/compose.h 1.1 1994/07/22 13:21:13 rex Exp $
@@ -45,10 +48,10 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 //	Prototypes
 
-void ComposeInit(grs_bitmap *pcompose, int bmtype, int w, int h);
+void ComposeInit(grs_bitmap *pcompose, int32_t bmtype, int32_t w, int32_t h);
 void ComposeAdd(grs_bitmap *pcompose, grs_bitmap *pbm);
-long ComposeDiff(grs_bitmap *pcompose, grs_bitmap *pbmNew, grs_bitmap *pbmDiff);
-long ComposeConvert(grs_bitmap *pcompose, grs_bitmap *pbm);
+int32_t ComposeDiff(grs_bitmap *pcompose, grs_bitmap *pbmNew, grs_bitmap *pbmDiff);
+int32_t ComposeConvert(grs_bitmap *pcompose, grs_bitmap *pbm);
 void ComposeFree(grs_bitmap *pcompose);
 
 //	Specific compose routines (type is compose buffer type, not bm!)
@@ -58,13 +61,13 @@ void ComposeFlat8Add(grs_bitmap *pcompose, grs_bitmap *pbm);
 
 //	Specific diff routines (type is compose buffer type, not bm!)
 
-long ComposeFlat8Diff(grs_bitmap *pcompose, grs_bitmap *pbmNew, grs_bitmap *pbmDiff);
-// long ComposeFlat24Diff(grs_bitmap *pcompose, grs_bitmap *pbmNew,
+int32_t ComposeFlat8Diff(grs_bitmap *pcompose, grs_bitmap *pbmNew, grs_bitmap *pbmDiff);
+// int32_t ComposeFlat24Diff(grs_bitmap *pcompose, grs_bitmap *pbmNew,
 //	grs_bitmap *pbmDiff);
 
 //	Specific convert routines (type is compose buffer type, not bm!)
 
-long ComposeFlat8Convert(grs_bitmap *pcompose, grs_bitmap *pbm);
-// long ComposeFlat24Convert(grs_bitmap *pcompose, grs_bitmap *pbm);
+int32_t ComposeFlat8Convert(grs_bitmap *pcompose, grs_bitmap *pbm);
+// int32_t ComposeFlat24Convert(grs_bitmap *pcompose, grs_bitmap *pbm);
 
 #endif
